requirements.txt
Pillow
tkinterdnd2
tkinterweb
markdown
